Now sittings at index show correct spotsleft

+++ sqlQueries/init.sql
@@ -112,7 +112,7 @@ create table sitting (
 	sittPayDeadline		date,
 	active				tinyint(1) DEFAULT 1,
 	resName				varchar(30),
-	spotsLeft			tinyint(1) DEFAULT 0,
+	spotsLeft			integer DEFAULT 0,
 	primary key(sittId),
 	foreign key(resName) references restaurant(resName) ON UPDATE CASCADE
 );
